Remove symlink
