--- server/src/protocol.ts
@@ -1,23 +1,0 @@
-const PacketTypeResize = 'resize';
-const PacketTypeData = 'data';
-
-// type PacketType = typeof PacketTypeResize | typeof PacketTypeData;
-
-interface ResizePacket {
-  packetType: typeof PacketTypeResize;
-  payload: {
-    cols: number;
-    rows: number;
-  };
-}
-
-interface DataPacket {
-  packetType: typeof PacketTypeData;
-  payload: string;
-}
-
-type Packet = ResizePacket | DataPacket;
-
-export {
-  Packet,
-}
