+++ tools/vscode-open.sh
