--- ui/src/protocol.ts
@@ -1,1 +1,0 @@
-../../server/src/protocol.ts
